bot/requirements.txt
discord.py==2.3.2
aiohttp==3.9.5
PyYAML==6.0.1
yt-dlp==2024.4.9
